lint code and style
changed the while test to NOT include an assignment. this is a code
style preference. while this change executes the regex test twice, for
loop test and variable assignment, execution times don’t change much.
BUT!!!!, this is Not efficient!! violates DRY. May change back and just
enter a comment above