insert check again for object/nsp.js
I am getting tired. need to stop coding. I want this stupid check. if a
variable is declared but not assigned a value, I want this function to
still perform. the variable is still undefined after the function is
run. so I will have to do this:

var oby
oby = nsp(oby, 'gail', 'apple')
console.log(oby)

for the variable to be meaningful…